[PM-26063] Move toolbar view extensions to BitwardenKit
